fix bug on points_types.h for type of intensity

+++ sdk/pandoraSDK/point_types.h
@@ -7,7 +7,7 @@ namespace pandar_pointcloud
 {
 struct PointXYZIT {
     PCL_ADD_POINT4D
-    uint8_t intensity;
+    float intensity;
     double timestamp;
     uint16_t ring;                      ///< laser ring number
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW // make sure our new allocators are aligned
@@ -17,7 +17,7 @@ struct PointXYZIT {
 
 POINT_CLOUD_REGISTER_POINT_STRUCT(pandar_pointcloud::PointXYZIT,
                                   (float, x, x)(float, y, y)(float, z, z)
-                                  (uint8_t, intensity, intensity)(double, timestamp, timestamp)(uint16_t, ring, ring))
+                                  (float, intensity, intensity)(double, timestamp, timestamp)(uint16_t, ring, ring))
 
 typedef pandar_pointcloud::PointXYZIT PPoint;
 typedef pcl::PointCloud<PPoint> PPointCloud;
